requirements.txt
streamlit
google-generativeai
python-dotenv
langchain-community
langchain
PyPDF2
chromadb
faiss-cpu
pytesseract
